pyproject.toml
[project]
name  = "thinking-mcp"
version="0.1.0"

description = "Thinking MCP"

authors = [{"name"  = "Mai Thành Duy An", "email"="tiachop0102@gmail.com"}]

requires-python= ">=3.12"

dependencies = [
    "requests",
    "fastapi",
    "uvicorn",
    "mcp",
    "PyJWT",
]


[project.scripts]
thinking-mcp="app.main:main"

[tool.setuptools]
packages = ["app"]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
